+++ R/build-articles.R
@@ -22,7 +22,7 @@
 #'
 #' @section Index and navbar:
 #' You can control the articles index and navbar with a `articles` field in
-#' your `_pkgdown.yaml`. If you use it, pkgdown will check that all articles
+#' your `_pkgdown.yml`. If you use it, pkgdown will check that all articles
 #' are included, and will error if you have missed any.
 #'
 #' The `articles` field defines a list of sections, each of which