+++ NEWS.md
@@ -1,3 +1,5 @@
+# pkgdown (development version)
+
 # pkgdown 2.0.5
 
 * Correctly generate downlit link targets for topics that have a file name